+++ packages/frontend/src/apis/queries/stocks/index.ts
@@ -1,5 +1,5 @@
 export * from './schema';
-export * from './useGetTopViews';
 export * from './useGetStocksByPrice';
 export * from './useGetSearchStocks';
 export * from './useGetStocksPriceSeries';
+export * from './useStockQueries';
