--- packages/frontend/src/apis/queries/stocks/useGetStockIndex.ts
@@ -1,17 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import { z } from 'zod';
-import { StockIndexSchema, type StockIndexResponse } from './schema';
-import { get } from '@/apis/utils/get';
-
-const getStockIndex = () =>
-  get<StockIndexResponse[]>({
-    schema: z.array(StockIndexSchema),
-    url: `/api/stock/index`,
-  });
-
-export const useGetStockIndex = () => {
-  return useQuery({
-    queryKey: ['stockIndex'],
-    queryFn: getStockIndex,
-  });
-};
